+++ benchmark/search.rb
@@ -28,7 +28,15 @@ if ENV["SETUP"]
     t.integer :store_id
   end
 
-  Product.import ["name", "color", "store_id"], total_docs.times.map { |i| ["Product #{i}", ["red", "blue"].sample, rand(10)] }
+  records = []
+  total_docs.times do |i|
+    records << {
+      name: "Product #{i}",
+      color: ["red", "blue"].sample,
+      store_id: rand(10)
+    }
+  end
+  Product.insert_all(records)
 
   puts "Imported"